Added code to run gprof and gcov
It should work to put the gcov info in the student log file by calling
the function I wrote for it. gprof isn't finished as I don't know where
to put its stats (e.g. student logfile, it's own file ... etc)

+++ main.cpp
@@ -57,7 +57,7 @@ int main(int argc, char ** argv)
 
 
         //loop through every .cpp and run it
-        for(i=0;i<cpps.size();i++)
+        for(i=0;i<(int)cpps.size();i++)
         {  
             //Excludes the "golden" .cpp from being evaluated 
             if(count(cpps.at(i).begin(), cpps.at(i).end(), '/') > 1)
